Login/logout

+++ app/components/appbar/AppBarUser.tsx
@@ -8,20 +8,37 @@ import Avatar from '@mui/material/Avatar';
 import Typography from '@mui/material/Typography';
 
 import { useRouter } from 'next/navigation';
-import React, { useEffect } from 'react';
-import { User } from '@/app/utils/fetch_module';
+import React, { useCallback, useEffect } from 'react';
+import * as fetch_mod from '@/app/utils/fetch_module';
+import { useCookies } from 'react-cookie';
 
-function AppBarUser(props: { user: User }) {
+function AppBarUser(props: { user: fetch_mod.User; authToken: string }) {
+	const [cookies, setCookie, removeCookie, updateCookies] = useCookies([
+		'auth-token',
+	]);
 	const router = useRouter();
+
+	const logout = () => {
+		fetch_mod.logout(props.authToken).then((was_successful) => {
+			if (was_successful) {
+				removeCookie('auth-token', {
+					expires: new Date(0),
+					path: '/',
+					sameSite: 'lax',
+				});
+				updateCookies();
+				purgeUserCache();
+				window.location.reload();
+			}
+		});
+	};
 
 	const PushToProfile = () => {
 		router.push('/profile?id=' + props.user.steamid);
 	};
 
 	const purgeUserCache = () => {
-		if (!confirm('This will log you out for a second. Continue?')) return;
 		window.localStorage.setItem('user-cache', 'null');
-		window.location.reload();
 	};
 
 	return (
@@ -43,13 +60,26 @@ function AppBarUser(props: { user: User }) {
 						<MenuItem onClick={PushToProfile}>Profile</MenuItem>
 						<MenuItem
 							onClick={() => {
-								purgeUserCache();
 								popupState.close();
+
+								if (!confirm('This will log you out for a second. Continue?'))
+									return;
+
+								purgeUserCache();
+								window.location.reload();
 							}}
 						>
 							Purge local user cache
 						</MenuItem>
-						<MenuItem onClick={popupState.close}>Logout</MenuItem>
+						<MenuItem
+							onClick={() => {
+								popupState.close();
+								if (!confirm('Are you sure you want to log out?')) return;
+								logout();
+							}}
+						>
+							Logout
+						</MenuItem>
 					</Menu>
 				</div>
 			)}
